pyproject.toml
[tool.poetry]
name = "smart-llm-loader"
version = "0.1.0"
description = "A powerful PDF processing toolkit that seamlessly integrates with LLMs for intelligent document chunking and RAG applications. Features smart context-aware segmentation, multi-LLM support, and optimized content extraction for enhanced RAG performance."
authors = ["drmingler <davidemmanuel75@gmail.com>"]
readme = "README.md"
packages = [{include = "smart_llm_loader"}]
license = "MIT"
repository = "https://github.com/drmingler/smart-llm-loader"
keywords = ["pdf", "llm", "rag", "document-processing", "ai"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Text Processing :: General"
]

[tool.poetry.dependencies]
python = "^3.9"
langchain = "^0.1.0"
langchain-community = "^0.0.10"
langchain-core = "^0.1.10"
requests = "^2.31.0"
python-dotenv = "^1.0.0"
pypdf = "^3.17.1"
faiss-cpu = "^1.7.4"
tiktoken = "^0.8.0"
litellm = "^1.61.3"
pdf2image = "^1.17.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.23.0"
pytest-cov = "^4.1.0"
pytest-mock = "^3.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=smart_llm_loader --cov-report=term-missing"
asyncio_mode = "auto"